Conflicts:
	bin/gui/admin/ModificarLibro.class
	bin/gui/user/ComprasRealizadas.class
	bin/model/Compra.class
	src/gui/admin/ModificarLibro.java

+++ src/model/Compra.java
@@ -15,7 +15,7 @@ public class Compra {
 	private Date fechaCompra;
 	private int numCuenta;
 	private int unidades;
-	private int importeTotal;
+	private double importeTotal;
 	
 	//Constructors
 	/**
@@ -85,11 +85,11 @@ public class Compra {
 		this.unidades = unidades;
 	}
 
-	public int getImporteTotal() {
+	public double getImporteTotal() {
 		return importeTotal;
 	}
 
-	public void setImporteTotal(int importeTotal) {
+	public void setImporteTotal(double importeTotal) {
 		this.importeTotal = importeTotal;
 	}
